requirements.txt
tweepy
kafka-python
vaderSentiment
pyspark
elasticsearch
